Typo: Remove Unnecessary `f` Format String Literal
This commit removes an unnecessary Python `f` format string literal in
the "02 Messages Format" lesson of the "API Fundamentals" course. It is
used twice for the prompt which asks Claude to "Generate a beautiful
haiku", however no Python variables are present within the prompt, and
so the `f` is not needed.